+++ lib/js2xml.ts
@@ -66,7 +66,8 @@ function writeAttributes(attributes: any, options: InternalOptions, depth: numbe
         if (Object.prototype.hasOwnProperty.call(attributes, key) && attributes[key] !== null && attributes[key] !== undefined) {
             quote = options.noQuotesForNativeAttributes && typeof attributes[key] !== 'string' ? '' : '"';
             attr = '' + attributes[key]; // ensure number and boolean are converted to String
-            attr = attr.replace(/"/g, '&quot;');
+            attr = attr.replace(/&amp;/g, '&'); // desanitize to avoid double sanitization
+            attr = attr.replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;');
             attrName = 'attributeNameFn' in options ? options.attributeNameFn(key, attr, currentElementName, currentElement) : key;
             result.push((options.spaces && options.indentAttributes ? writeIndentation(options, depth + 1, false) : ' '));
             result.push(attrName + '=' + quote + ('attributeValueFn' in options ? options.attributeValueFn(attr, key, currentElementName, currentElement) : attr) + quote);